Help now recommends using online web page
Since the remote aczoom.com web page can be used by everyone for accessing
TVs in their home networks using a browser on their home networks,
simplified the help to recommend that as the easiest method.

Also reduced the size of the Vol and CH buttons to allow them to be
displayed in a single row on smartphone screens.