Use repeated setTimeout calls instead of one call
setTimeout does not look at wall clock time, so if the browser tab for
this page is paused by moving to another tab in the browser, when the user
comes back to this tab the setTimeout will continue to keep waiting,
ignoring all the time this tab was not in front.

To avoid having the popup display message stay up for too long, the code
now uses setTimeout for displayPopup for a short period, then looks at
wall clock time, and repeats the setTimeout if needed, until the required
wall clock time is reached.