Add support for Sony TV REST API JSON RPC
In addition to IRCC-IP (which supports all remote control commands), now
this supports sending JSON messages to the Sony TV using the REST API.

A new tab REST API created. It has one command getApplicationList
hardcoded, and has a form where users can select the service name and
enter JSON strings to describe the method to call and the parameters to
send for that call.
The Sony reference documentation has examples of all the API commands
and the JSON messages they require:
https://pro-bravia.sony.net/develop/integrate/rest-api/spec/index.html